change type to double

+++ backend/src/main/java/org/codecool/fitnesstracker/fitnesstracker/controller/dto/ActivityCalorieForAnalyticsDTO.java
@@ -2,5 +2,5 @@ package org.codecool.fitnesstracker.fitnesstracker.controller.dto;
 
 import java.time.LocalDateTime;
 
-public record ActivityCalorieForAnalyticsDTO(int calorie, LocalDateTime activityTime) {
+public record ActivityCalorieForAnalyticsDTO(double calorie, LocalDateTime activityTime) {
 }
